+++ src/module_shims/audio-vs1053b.h
@@ -72,9 +72,9 @@
 #define OUTPUT_GPIO 7
 
 // Play a buffer. Will stop any currently playing buffer first.
-int8_t audio_play_buffer(uint8_t command_select, uint8_t data_select, uint8_t dreq, const uint8_t *buf, uint32_t buf_len);
+int32_t audio_play_buffer(uint8_t command_select, uint8_t data_select, uint8_t dreq, const uint8_t *buf, uint32_t buf_len);
 // The same as play except it will append the buffer to a currently playing buffer if there is one (used for streams)
-int8_t audio_queue_buffer(uint8_t command_select, uint8_t data_select, uint8_t dreq, const uint8_t *buf, uint32_t buf_len);
+int32_t audio_queue_buffer(uint8_t command_select, uint8_t data_select, uint8_t dreq, const uint8_t *buf, uint32_t buf_len);
 // Stops any audio playing and frees memory
 int8_t audio_stop_buffer(bool immediately);
 // Stops music playback
